refactor(demos): migrate persistence cluster (imperative-persistence, storage-shape, edge-cases)
Three persistence demos move flat -> folder onto the shared registry.

imperative-persistence keeps its dark terminal append-log (a true
console reads the same in light and dark) and maps its mono command
buttons onto .actions.mono. storage-shape's two JSON readouts collapse
onto the light pre per the canonical-readout policy (dark boxes are
reserved for real append-logs, not state dumps), with its empty-state
box on .placeholder. persistence-edge-cases collapses its bespoke
.tag-row onto the canonical field-array .row, keeping only the one-line
filling-input residue that the committed array demos already share.

Folds out em-dashes from the log format, the imperative record() copy,
and the edge-cases field captions while the files are open (commas,
colons, and bracketed timestamps in their place).
