+++ apps/site/docs-demos/persistence-edge-cases/styles.json
@@ -1,0 +1,3 @@
+{
+  "with": ["label", "input", "button", "fieldset", "row", "ghost", "hint", "code"]
+}
